requirements.txt
about-time==4.2.1
absl-py==1.4.0
alive-progress==3.1.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
autograd==1.5
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
cachetools==5.3.0
celluloid==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cma==3.2.2
cmake==3.26.3
comm==0.1.4
contourpy==1.0.7
cvxpy==1.3.1
cycler==0.11.0
Cython==0.29.35
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
docker-pycreds==0.4.0
ecos==2.0.12
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.18
fastjsonschema==2.18.1
filelock==3.12.0
fonttools==4.39.3
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grapheme==0.6.0
grpcio==1.54.0
#gym==0.21.0
hydra-core==1.2.0
idna==3.4
imageio==2.31.1
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.4
lit==16.0.2
lockfile==0.12.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
mujoco==2.3.6
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
osqp==0.6.2.post9
overrides==7.4.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
-e git+https://github.com/benelot/pybullet-gym.git@bc68201c8101c4e30dde95f425647a0709ee2f29#egg=pybulletgym
pycodestyle==2.10.0
pycparser==2.21
pyglet==1.5.27
Pygments==2.16.1
pymoo==0.6.0.1
PyOpenGL==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
qdldl==0.1.7
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.21.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.0
tornado==6.3.3
traitlets==5.10.1
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
wandb==0.15.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.0
wrapt==1.15.0
